rename and add warnings
