--- packages/core/src/constants/function-literal-prefix.ts
@@ -1,1 +1,0 @@
-export const functionLiteralPrefix = `@builder.io/mitosis/function:`;
