--- packages/core/src/constants/method-literal-prefix.ts
@@ -1,1 +1,0 @@
-export const methodLiteralPrefix = `@builder.io/mitosis/method:`;
