+++ packages/core/src/parsers/builder.ts
@@ -13,7 +13,7 @@ import { MitosisNode } from '../types/mitosis-node';
 import { parseJsx, parseStateObject } from './jsx';
 import { parseCode, isExpression } from '../helpers/parsers';
 import { hashCodeAsString } from '..';
-import { mapJsonObjectToStateValue } from '../helpers/state';
+import { mapJsonObjectToStateValue } from './helpers/state';
 import { JSONObject } from '../types/json';
 
 // Omit some superflous styles that can come from Builder's web importer
