+++ packages/core/src/parsers/constants/outdated-prefixes.ts
@@ -1,0 +1,10 @@
+/**
+ * DO NOT USE THIS PREFIX. It's only used for legacy purposes, within a very small subset of logic in the JSX & Builder parsers.
+ * It will be eliminated in the future, and generators should not need to rely on these prefixes for any reason.
+ */
+export const __DO_NOT_USE_FUNCTION_LITERAL_PREFIX = `@builder.io/mitosis/function:`;
+/**
+ * DO NOT USE THIS PREFIX. It's only used for legacy purposes, within a very small subset of logic in the JSX & Builder parsers.
+ * It will be eliminated in the future, and generators should not need to rely on these prefixes for any reason.
+ */
+export const __DO_NOT_USE_METHOD_LITERAL_PREFIX = `@builder.io/mitosis/method:`;
